chore(api): upgrade test toolchain to vitest 4 + pool-workers 0.16
Upgrades @cloudflare/vitest-pool-workers 0.5→0.16.20, vitest 2→4.1.9, and
wrangler→4.105.0 to pull in the Windows workerd loopback-socket stability
fixes. This is the real fix for the intermittent pre-push test flake
(ConnectEx #1225/#52, ECONNREFUSED 127.0.0.1) that the lefthook retry was
papering over; the retry stopgap is removed.

Migration changes required by the new toolchain:

- Config: defineWorkersConfig + poolOptions.workers → the cloudflareTest()
  plugin inside defineConfig. The package is ESM-only, so the config is now
  vitest.config.mts.
- Storage isolation is per-file (the per-test isolatedStorage flag was
  removed). Give seedWorkspace a unique default slug and clear the D1
  rate-limit counter in a global beforeEach so transient state no longer
  bleeds across tests in a file.
- Filter the spurious "unhandled rejection" reports workerd raises for the
  ServiceErrors that MCP tool handlers throw and routes/mcp.ts catches, via
  a scoped test.onUnhandledError (only our error kinds; real leaks still
  fail).
- cloudflare:test ambient types moved to the /types subpath; point tsconfig
  there and augment Cloudflare.Env (ProvidedEnv was removed) from the
  project Env.
