+++ apps/api/package.json
@@ -21,11 +21,11 @@
     "zod": "^4.4.3"
   },
   "devDependencies": {
-    "@cloudflare/vitest-pool-workers": "^0.5.0",
+    "@cloudflare/vitest-pool-workers": "^0.16.20",
     "@cloudflare/workers-types": "^4.20240924.0",
     "tsx": "^4.19.0",
     "typescript": "^5.5.0",
-    "vitest": "^2.1.0",
-    "wrangler": "^4.0.0"
+    "vitest": "^4.1.9",
+    "wrangler": "^4.105.0"
   }
 }
