+++ apps/api/src/test/env.d.ts
@@ -1,5 +1,10 @@
-import type { Env } from "@projektor/types";
+import type { Env as ProjektorEnv } from "@projektor/types";
 
-declare module "cloudflare:test" {
-	interface ProvidedEnv extends Env {}
+// vitest-pool-workers ≥0.13 types `cloudflare:test`'s `env` as `Cloudflare.Env`
+// (the old `ProvidedEnv` interface was removed). Augment that global namespace
+// with our binding shape so `env.DB` / `env.KV` / … are typed in tests.
+declare global {
+	namespace Cloudflare {
+		interface Env extends ProjektorEnv {}
+	}
 }
